Consolidate pricing data to it's own file
