--- src/components/Pricing/interfaces.ts
@@ -1,12 +1,0 @@
-/**
- * The features to display for pricing plans.
- */
-export interface FeatureCluster {
-    topic: string
-    features?: string[]
-}
-
-export interface FeatureInfo {
-    label: string
-    description?: string
-}
